commandHauls: extract Command to module
